GH-1593: Moved product-config
