+++ projects/core/src/product/config/product-config.ts
