+++ projects/core/src/product/occ/index.ts
@@ -1,4 +1,3 @@
 export * from './product.service';
 export * from './product-search.service';
 export * from './product-reviews.service';
-export * from './product-config';
